Исправления по ревью 1
