+++ api_yamdb/api/serializers.py
@@ -11,52 +11,10 @@ from .serializer_fields import RatingByScoresField
 from reviews.models import Category, Title, Genre, User, Review, Comment
 
 
-class UsersSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = (
-            'username',
-            'email',
-            'first_name',
-            'last_name',
-            'bio',
-            'role',
-        )
 
 
-class NoAdminSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = (
-            'username',
-            'email',
-            'first_name',
-            'last_name',
-            'bio',
-            'role',
-        )
-        read_only_fields = ('role',)
 
 
-class TokenSerializer(serializers.ModelSerializer):
-    username = serializers.CharField(required=True)
-    confirmation_code = serializers.CharField(required=True)
-
-    class Meta:
-        model = User
-        fields = ('username', 'confirmation_code')
-
-    def validate(self, data):
-        user = get_object_or_404(User, username=data.get('username'))
-        if user.confirmation_code != data.get('confirmation_code'):
-            raise serializers.ValidationError('Не верный confirmation_code')
-        return {'access': str(AccessToken.for_user(user))}
-
-
-class SignupSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('email', 'username')
 
 
 class CategorySerializer(serializers.ModelSerializer):
