+++ api_yamdb/api/utils.py
@@ -1,5 +1,5 @@
 from random import choices
 
 
-def generate_confirmation_code():
-    return ''.join(choices('123456789', k=4))
+def generate_confirmation_code(length=4):
+    return ''.join(choices('0123456789', k=length))
