+++ api_yamdb/custom_user/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from custom_user.models import User
+
+admin.site.register(User)
