+++ api_yamdb/custom_user/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CustomUserConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'custom_user'
