+++ api_yamdb/reviews/models.py
@@ -6,8 +6,8 @@ from django.core.validators import (
 )
 from django.contrib.auth.models import AbstractUser
 
-from .validators import validate_username
-from .constants import USER, ADMIN, MODERATOR, MAX_LENGTH, USERS_ROLE
+from custom_user.models import User
+
 
 
 class Category(models.Model):
@@ -44,35 +44,7 @@ class GenreTitle(models.Model):
     genre = models.ForeignKey(Genre, on_delete=models.CASCADE)
 
 
-class User(AbstractUser):
-    username = models.CharField(
-        validators=(validate_username, RegexValidator(regex=r'^[\w.@+-]+\Z')),
-        max_length=MAX_LENGTH,
-        unique=True,
-    )
-    email = models.EmailField(max_length=254, unique=True, null=False)
-    role = models.CharField(
-        max_length=20, choices=USERS_ROLE, default=USER, blank=True
-    )
-    bio = models.TextField(
-        blank=True,
-    )
-    confirmation_code = models.CharField(
-        max_length=MAX_LENGTH,
-        blank=True,
-    )
 
-    @property
-    def is_user(self):
-        return self.role == USER
-
-    @property
-    def is_admin(self):
-        return self.role == ADMIN
-
-    @property
-    def is_moderator(self):
-        return self.role == MODERATOR
 
 
 class Review(models.Model):
