--- api_yamdb/reviews/validators.py
@@ -1,8 +1,0 @@
-from django.core.exceptions import ValidationError
-
-
-def validate_username(value):
-    if value == 'me':
-        raise ValidationError(
-            'Использовать имя "me" в качестве username запрещено.',
-        )
